Hent pdl søker og barn. Map til søkerinfo
