+++ src/main/kotlin/no/nav/familie/ef/søknad/service/OppslagService.kt
@@ -8,6 +8,5 @@ interface OppslagService {
 
     fun hentSøkerinfoV2(): Søkerinfo
 
-    fun hentSøkerinfoV3(): Søkerinfo
 
 }
